+++ lib/gui/qrListAppBar.dart
